[WIP] Migrate away from using ResourceIdentifier now that it's gone upstream
As part of upstream bug 1134547[1], the ResourceIdentifier that Brave
Shields relied on so far has been completely removed[2], so we need to
adapt our code not to rely on it anymore so that we can keep building
Brave correctly on the latest dev channel release of Chromium 88.

This patch implements the first step to adapt Brave to this new
scenario, by moving into using Brave-specific values of the
ContentSettingsType enumeration instead of using the PLUGINS
type plus a ResourceIdentifier as we did before.

Other important changes done here:

  - Brave-specific content types are registered with PERSISTENT storage.
  - Removed BraveEphemeralProvider, no longer needed due to using now
    Brave-specific content types with persistent storage.
  - Added a "bs_" prefix to all the BraveShields constants, to avoid
    collisions with the COOKIES and JAVASCRIPT upstream content types.
    Added prefix to all constants for consistency + update TS code.
  - Stop re-registering the PLUGINS content setting type to block it
    by default. No needed as we use Brave-specific content types now.

Note that the goal of this change is simply to keep Brave building,
running and passing the tests so that we can continue working on the
rebase. Further work will be still needed in follow-up patches to
fully migrate the code base to a post-Flash world by changing those
bits that don't make sense anymore, and simplifying things as much
as possible.

[1] https://crbug.com/1134547
[2] https://crrev.com/c/2500161